RelationInputWrapper: Fix prop-types, add mock data
