+++ packages/core/admin/admin/src/content-manager/components/RelationInputWrapper/utils/getRelationLink.js
@@ -1,5 +1,5 @@
 import { getRequestUrl } from '../../../utils';
 
 export function getRelationLink(targetModel, id) {
-  return getRequestUrl(`collectionType/${targetModel}/${id ?? ''}`);
+  return `/admin${getRequestUrl(`collectionType/${targetModel}/${id ?? ''}`)}`;
 }
