+++ packages/core/admin/admin/src/content-manager/components/RelationInputWrapper/utils/tests/getRelationLink.test.js
@@ -2,10 +2,12 @@ import { getRelationLink } from '../getRelationLink';
 
 describe('getRelationLink', () => {
   test('returns an URL containing the targetModel and id', () => {
-    expect(getRelationLink('model', 2)).toBe('/content-manager/collectionType/model/2');
+    expect(getRelationLink('model', 2)).toBe('/admin/content-manager/collectionType/model/2');
   });
 
   test('returns an URL containing the targetModel', () => {
-    expect(getRelationLink('model', undefined)).toBe('/content-manager/collectionType/model/');
+    expect(getRelationLink('model', undefined)).toBe(
+      '/admin/content-manager/collectionType/model/'
+    );
   });
 });
